Major refactoring with unusable middle steps -> no commit earlier... This compiles but only the KSaneDevice works somehow not KSaneWidget.
Changes:
- Remove the namespace KSaneIface and make sure all classes start with KSane as
they mostly have done already.

- Remove the creategui() function KSaneOption -> gui-less
- Rename Class files to the same name to simplify #include "..."
- Rename header protections to be the Class name + "_h" for easier class renames
- Rework authentication and sane access (only one sane_init() / process)
- Replace float with qreal
- And more...

+++ ksanetwain/ktwain_widget.cpp
@@ -138,8 +138,8 @@ bool KSaneWidget::setOptVal(const QString &option, const QString &value) {return
 void KSaneWidget::setScanButtonText(const QString &scanLabel) {}
 void KSaneWidget::setPreviewButtonText(const QString &previewLabel) {}
 void KSaneWidget::enableAutoSelect(bool enable) {}
-float KSaneWidget::currentDPI() {return 0.0;}
-void KSaneWidget::setPreviewResolution(float dpi){Q_UNUSED(dpi);}
+qreal KSaneWidget::currentDPI() {return 0.0;}
+void KSaneWidget::setPreviewResolution(qreal dpi){Q_UNUSED(dpi);}
 void KSaneWidget::setSelection(QPointF lefttop, QPointF rightbottom){Q_UNUSED(lefttop); Q_UNUSED(rightbottom);}
 void KSaneWidget::setOptionsCollapsed(bool){}
 void KSaneWidget::setScanButtonHidden(bool){}
