+++ libksane/widgets/labeled_fslider.h
@@ -62,21 +62,21 @@ public:
     * \param quant is the step between values.
     */
     LabeledFSlider(QWidget *parent, const QString& text,
-                   float min, float max, float st);
+                   qreal min, qreal max, qreal st);
     ~LabeledFSlider();
 
    /**
     * \return the slider value.
     */
-    float value() const {return (float)m_spinb->value();}
-    float step() const {return m_fstep;}
+    qreal value() const {return (qreal)m_spinb->value();}
+    qreal step() const {return m_fstep;}
 
 public Q_SLOTS:
 
     /** Set the slider/spinbox value */
-    void setValue(float);
-    void setRange(float, float);
-    void setStep(float);
+    void setValue(qreal);
+    void setRange(qreal, qreal);
+    void setStep(qreal);
     /** Set the unit */
     void setSuffix(const QString &text);
 
@@ -95,13 +95,13 @@ Q_SIGNALS:
     /**
      * Emit the slider value changes
      */
-    void valueChanged(float);
+    void valueChanged(qreal);
 
 private:
 
     QSlider        *m_slider;
     QDoubleSpinBox *m_spinb;
-    float           m_fstep;
+    qreal           m_fstep;
     int             m_istep;
 };
